#440: Move nested attribute unit tests to own class.
